Convert memory_exhaust.php to .hhas
Summary: The memory_exhaust.php test is trying to exercise a pathologically sized edge case in the assembler, but it first has to make it through the parser/emitter. Since we have other tests checking that the parser fails gracefully for overly large expressions, turn this one into .hhas, bypassing the parser/emitter.

Reviewed By: oulgen

Differential Revision: D15257472

fbshipit-source-id: b387155e5ba7cd2ef6fc57e3f0aec5cae14ed53d