Keep track of origin of changes for SearchServiceRunner
Summary:
In a future diff in this stack, we will need to know whether a file enqueued for the symbol index came from A) the saved state, or B) the typechecker.

This diff records `file_change_detector` as a third element in the `SearchServiceRunner.queue`.

In future diffs, some symbol indexes will parse changes differently if they came from the saved state vs coming from the typechecker.

Reviewed By: 2BitSalute

Differential Revision: D15228813

fbshipit-source-id: f4206713b0dd4fe5eacaddf8ea95fcb2babe81ab