Don't check the autoload map for reified functions
Summary:
Historically, HHVM has dealt with generic types through type erasure, by replacing types like

```
Ref<dict<classname<MeerkatStep>, Time>>
```

with the simpler type name `Ref`. So generics are essentially only part of the typechecker, with no bearing on any part of the runtime. This cuts out a bunch of information which we could be using to enforce soundness and generate performant code.

With reified generics, we instead create new mangled type names based on the generic parameters. Here, we treat this type inside the runtime as

```
$$Ref$$<HH\dict<string,Time>>
```

These types and functions don't exist in the autoload map. Unfortunately we've been asking the autoloader to look for them anyway, even though we should know better. This diff checks whether a given function name is reified a little earlier in the flow.

We now call `isReifiedName` once for every unreified function we encounter, but we no longer call `AutoloadHandler::autoloadFunc` for every reified function we encounter. Hopefully performance-wise this will be a wash.

Reviewed By: oulgen

Differential Revision: D15289091

fbshipit-source-id: 4f9fff42fd43960a14fb0d85d40d22c56e7018b0