Make ty <: ?ty hold if ty is an expression dependent type
Summary:
We expect `ty <: ?ty` hold universally, for all types `ty`.  Astonishingly, as currently implemented, the subtyping relationship does not satisfy this property if `ty` is an expression dependent type.  Here is an example (added as a test case below):
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

class C {
  public function get(): this {
    return $this;
  }

  public function set(?this $_): void {}
}

function test(C $c): void {
  $d = $c->get();
  $d->set($d);
}
```
The type of `$d` is `<expr#1> as C`, and in order for the call `$d->set($d)` to be well-typed `<expr#1> as C` must be a subtype of `?(<expr#1> as C)` (`this` resolves to `<expr#1> as C` for `$d`).  However, we simplify `<expr#1> as C <: ?ty_super` to `C <: ?ty_super`, i.e., we only try to satisfy the relationship through the upper bound.  In this case it cannot be satisfied that way, but it can be satisfied through `ty_super`, i.e., `<expr#1> as C <: ty_super <: ?ty_super`.  Notice that we do exactly this for dependent types (we simplify the subtyping constraint to a disjunction of simpler constraints) but have a special case for dependent types of the form ``Tabstract (AKdependent (`expr _, []), Some ty)``.  I don't understand why, and the above example shows that this is not enough.  (It is not unsound but we reject valid programs.)  I'm proposing to put this case back.

Reviewed By: andrewjkennedy

Differential Revision: D15279318

fbshipit-source-id: 89d7a3489d137a05c589d868d493f8bd204affae