Fix check argument heuristic
Summary:
When deciding whether to try to get a context sensitive return type,
we look at the supplied arguments, and check to see if at least one of
them is better than we could deduce from the type constraint (in the
default configeration, we only consider functions that return one of
their parameters, and we only check the parameter that it function
returns).

But the test is wrong; if its not more refined, it returns true,
otherwise, if its not equivalently refined, it returns true.

This effectively means we consider *any* type that isnt exactly the
same as the constraint to be worth investigating.

I don't expect fixing this to make any difference to the final results
(we were trying cases that simply couldn't produce better results than
the non-context-sensitive return type); but it might improve
performance a little. It *might* even make it possible to enable it
for all functions (rather than just functions that return one of their
parameters).

Reviewed By: mofarrell

Differential Revision: D15284493

fbshipit-source-id: d7b20824928b5406d8f3ed1b492ecce23a4cdf52