Avoid calling packedData within packed-array.cpp
Summary:
This code replaces most usages of packedData throughout our codebase with calls to GetValueRef (for const access) and LvalUncheckedInt (for mutable access).

In terms of performance, I think the main potential issue is replacing pointer-based iteration with index-based iteration in several places, but canaries don't show an effect. It's possible that we'll want to replace some of the manual iteration with calls to IterateV / IterateKV in a follow-up - I'm still experimenting with those helpers.

I caught an infinite-loop bug in my first version of CopyPackedHelper. The bug occurs when converting an array containing a ref into a vec. I can't figure out how to set up this situation in a test.

Reviewed By: mxw

Differential Revision: D15070781

fbshipit-source-id: 876d5f8a4ec37a90a85df0d18416e1a704bfbdba