Use NoOverride flag on consts to optimize ClsCns
Summary:
This diff adds a NoOverride flag to Const data structure in HHBBC that is set when there are no overrides to this constant in any classes that extend the current class or through interfaces etc.

On this diff, this NoOverride flag is used to optimize `ClsCns` when it is used in `static::C` where `C` is a constant on a class.

On the next diff, I'll be using this NoOverride flag to optimize `type_structure(static::class, cnst)`.

Reviewed By: markw65

Differential Revision: D15228716

fbshipit-source-id: 445f65f5e94e79f1d5b985012edf2495c9fc38d2