Record cold sections of unique stubs
Summary:
When generating unique stubs, we record the code ranges for
various things like GDB and the perf pid map. However we were only
recording the range in the main code block. So, any part of the stub
in a cold/frozen block would not be recorded. We can use TransLoc to
record all the blocks. It looks like some of the prologue dispatch
code had similar issues, so fix that there too.

Reviewed By: ottoni, paulbiss

Differential Revision: D15282150

fbshipit-source-id: 1925cd0c908b340bac0ef32ac868cefabfee649c