Update dune version + add merlin and ocp-indent in the mini-repo
Summary:
The current OSS build doesn't deal correctly with the injector_config module
and links the default_injector everywhere. This prevents us to execute
a large part of the integration test suite. By bumping dune's version
to 1.9.3 we are getting the opportunity to use virtual libraries which
are a very elegant solution to this issue.

Also, in order to provide OSS devs with the best experience, I added merlin
and ocp-indent (and deps) to the mini-repo. I also bumped dune's version
to a recent one in order to be able to use virtual libraries.

Reviewed By: fredemmott

Differential Revision: D15294554

fbshipit-source-id: 831a1abab5028d19cc21b9731a56609721408aef