New inference: fix --remove-dead-fixmes for TAST check errors
Summary:
The `--remove-dead-fixme` option does not work under `new_inference` for the array index type errors caught by `invalid_arraykey_check.ml`: it removes `HH_FIXME[4110]` for that error, wrongly.

The problem lies with the use of `Errors.ignore_` and `Errors.try_` wrapping a subtype assertion in `tast_env.ml`. There is a horrible hack in the equivalent function (`is_sub_type`) in `Typing.subtype.ml` to patch back the `HH_FIXME` list. So let's use that function instead of re-implementing it incorrectly.

Reviewed By: manzyuk

Differential Revision: D15316813

fbshipit-source-id: e9a35ca8953673b78f69dd1dea1cf48b87b13de5