Include instantiation of generic types in error messages
Summary:
Error messages currently use the phrase "an object of type C" even for generic types. Better to be precise, and state "an object of type C<int,string>".

The pretty-printer code is reorganized a bit. In particular, the signature constrains it to operate on `locl ty` but the implementation works on arbitrary types, redundantly. So it was possible to remove quite a bit of code.

Reviewed By: Wilfred, CatherineGasnier

Differential Revision: D15294837

fbshipit-source-id: 1e90083ba3de50a44cba4c17b0c481a0c7e4ba61