Emit FCall instead of side-exit when profiled callee is not strongly biased
Summary:
This avoids the cost of side-exiting the region when the profile-driven function
check fails, which then allows more aggressively inserting those checks.  This
diff also takes advantage of these opportunities by lowering the threshold for
inserting those checks.  When the profile is strongly biased towads a single
callee, we still keep the side exit, as it can potentially lead to more
aggressive optimizations by avoiding the merge point after the inlined return.

Reviewed By: ricklavoie

Differential Revision: D15311747

fbshipit-source-id: 809d32a32898a9d921ceb18eacb69901f9021d12