Look up synthesized ancestors lazily
Summary:
Replace the use of tc_req_ancestors_extends (the set of names of all require-extends and require-implements ancestors, plus the names of all classes they extend, traits they use, and XHP classes whose attributes they include) when shallow_class_decl is enabled using the linearization.

To do this, we need only collect the names of all ancestors marked with the mro_synthesized flag in the linearization.

Reviewed By: vassilmladenov

Differential Revision: D14631449

fbshipit-source-id: 3c914fec54fe08cc1ae57b08fc40729406d6dd4d