Replace tc_req_ancestors with lazy-lookup
Summary:
Replace the use of tc_req_ancestors (the list of all require-extends and require-implements constraints applied to a given class) with lazy lookup from the linearization when shallow_class_decl is enabled.

To do this, we need to mark MRO elements when they were included via the ReqExtends or ReqImpl source_type (mro_synthesized is not sufficient, since *ancestors* of a require-extends class are also marked as synthesized). To emit a useful error, we also need to preserve the position at which an ancestor requirement was most recently included in the hierarchy (We could make the error message even better by using a *stack* of positions, but I'm only doing what's necessary here to preserve the behavior of legacy decl for now). The new field mro_required_at serves both of these purposes. When it is Some, the mro_element represents an ancestor requirement (a class which appeared in a require-extends or require-implements clause), and the position represents the location where that requirement was most recently included in the linearization (i.e., always a position within the declaration of the class which was linearized, indicating which of its direct ancestors involved the requirement).

Reviewed By: vassilmladenov

Differential Revision: D14631448

fbshipit-source-id: 7ad186b8377cb134d861464a0908d238a3fd5fb9