Add command to retrieve a list of dependent files
Summary:
Add a command to hh_client to find potentially dependent files given a list of files.

```
$ hh_client --file-dependents foo.php path/to/bar.php ...
qux.php
baz.php
...
```

Given a set of filepaths, find the set of files in the codebase that have a typing dependency on the input set.

foo.php: declares `class Foo {}`
bar.php: `class Bar extends Foo {}`
baz.php: `function baz(Bar $_): void {}`
qux.php: `function qux(): void { baz(); }`

bar.php has a typing dependency on foo.php
baz.php has a typing dependency on foo.php and bar.php.
quz.php has a typing dependency on file baz.php but not on foo.php or bar.php.

This command prints neither a subset nor a superset (or an exact set) of the actual file dependencies. Because there are still dynamic and unsound parts of Hack and the Hack codebase, it is possible that there are runtime dependencies that are not recorded for typechecking and would not be reported in this command. (Thus, it is not an exact or superset). This command can also print dependencies that are not part of the actual set because of the way that hh_server records dependencies. If file B has a dependency on file A and then you remove the dependency between file B and file A, file B will still have a potential dependency in the graph and will be reported (until hh_server is restarted). Thus, not necessarily a strict subset of runtime dependencies either.

Differential Revision: D15272083

fbshipit-source-id: 9f170d0cdd2423d21be1b03108964df8a719afae