Use virtual library to correctly link injector_config
Summary:
injector_config is now declared as a virtual library (in short, a .mli only)
which can be used by other libraries for compilation. One must use the
`injector_config` library in other libraries.
There are two implementations available, `default_injector` and
`test_injector`, which both declare that they `implement injector_config`.
When building an executable, one must use an implementation (and exactly one).
It doesn't have a big impact on the binaries itself, but we are going to
be able to add more tests to the build

Reviewed By: Wilfred

Differential Revision: D15317152

fbshipit-source-id: 6fe0ee8ac599399eb7188027c0f3c5e223a0ce80