bug in Typing_union.diff
Summary:
Typing_union.diff is a simplistic implementation of the type difference which assumes ty1 is a normalized union and ty2 is a "bare" type (no Toption or Tunresolved). It could however be enhanced by enforcing:
- T \ T = nothing
- ?T \ T = null

Those cases would have been caught up by Typing_unify_recursive.add anyway so those changes are not visible for now, but will be in a later diff.

Reviewed By: manzyuk

Differential Revision: D15319462

fbshipit-source-id: bcce64b7aca334338e9a17f939761ee08bf2b8b1