Better DCE for catch traces
Summary:
Values stored to the stack in normal catch traces are going to be
popped by the unwinder. If the value being stored is IncReffed in the
catch trace, or was an otherwise dead PRc instruction, we can kill the
IncRef/PRc instruction, and replace the source of the store with a
constant InitNull (a follow up diff will attempt to adjust the stack
to avoid writing anything).

This will reduce the lifetime of the temporaries, and typically avoid
forcing them into callee save registers, or spilling them to the
stack. It can also kill instructions that are otherwise unused on the
main trace.

Reviewed By: paulbiss

Differential Revision: D15314371

fbshipit-source-id: 4748a777b8bb570de6f3e5762f87b834c1e3b38c