Add mode to EndCatch
Summary:
Differentiate catch traces that might be used for VMSwitchMode,
VMBuiltinSwitchMode, or special case handling via side exit.

For "normal" catch traces, values being spilt to the stack are only
there for the unwinder to pop them. Other kinds of catch traces might
actually consume the stack slots, so in order to optimize away such
stores, we need to know the kind of catch trace we're dealing with.

This diff doesn't really use the mode. I'll add optimizations in later
diffs.

Reviewed By: paulbiss

Differential Revision: D15296707

fbshipit-source-id: 5d8a68065aa7549755c9ee2bd47e3910fae1750f