Implement local search service
Summary:
Now that autocomplete data can be kept in a sqlite file that is loaded from a saved-state, we need a way to keep track of locally changed files since that saved state.

This implementation is the simplest possible form of tracking for locally changed `fileinfo.t` and `fileinfo.names` objects.

* It does not rewrite these `fileinfo` objects into a different data structure, so as to avoid copying data and consuming more memory.
* It avoids tracking anything captured from the naming table saved-state, and only tracks things touched by the typechecker.
* It does very simple string comparisons to find matches.

As long as the size of this data structure does not grow beyond a few thousand files, string comparisons are probably fast enough.

If we find that in the real world we are capturing more than a few thousand files in the local tracker, this should be rewritten.  It might be possible to replace this with reuse of the TrieIndex, or we could potentially just write changes directly to sqlite.

Note that this local search provider does not capture **"symbols removed"**. If you remove a symbol locally that exists in the global symbol file, the symbol will continue to appear in the results.  Note the discussion below.

Reviewed By: ljw1004

Differential Revision: D15281375

fbshipit-source-id: 0451505cb8da8123fa7452bce5052fe9add04d63