Static arrprov
Summary:
Add another side-table w/ thread-safe access to store provenance info for static arrays--also make sure we don't merge two static adatas if they have distinct provenance tags--basically, we have to bake the provenance info into the hash.

The provenance info is calculated based on the srcloc where we reference the adata in the output unit--this means that we need to ensure hackc never generates a unit where one adata is used in more than one place, which will be fixed in a stacked diff.

There's some trouble regarding needing to delay the allocation of the array until we know the srcloc for its provenance tag to avoid triggering an OOM since we can't make it uncounted until we add the provenance tag in, since it goes into the array hash.

This will be ruined by HHBBC at present and arrays pulled out of a repo won't have provenance tags on them. Both will be fixed in stacked diffs.

Reviewed By: mxw

Differential Revision: D14881020

fbshipit-source-id: e8dabeacdefd6c0a849da581e74a1f88f340982a