Improve error messages produced for shape indexing functions
Summary:
To type check a call to an indexing function on shapes, `Shapes::idx` or `Shapes::at`, we create a "fake" shape type of the form `shape(?sfn => v, ...)`, where `sfn` is the name of the field being looked up and `v` is a fresh type variable, and assert that the type of the shape argument must be a subtype of that fake shape type.  At the moment, we don't attach any reason information to the fake shape type, which occasionally leads to errors containing no source position.  E.g.,
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

type Shape = shape(?'x' => ?shape('y' => int));

function test(Shape $s): void {
  Shapes::idx(
    /* HH_FIXME[4110] */
    Shapes::idx($s, 'x', shape()),
    'y',
    0,
  );
}
```
would produce the following confusing error message under the new inference:
```
File "/home/manzyuk/fbsource/fbcode/hphp/hack/test/typecheck/nested_shapes_idx_hh_fixme.php", line 10, characters 5-7:
Invalid argument (Typing[4110])
File "", line 0, characters 0-0:
This is a shape
File "/home/manzyuk/fbsource/fbcode/hphp/hack/test/typecheck/nested_shapes_idx_hh_fixme.php", line 4, characters 28-45:
It is incompatible with a nullable type
```
The `HH_FIXME` is necessary to trigger this particular error because it hides another error, detected earlier in the process of type checking `test`, that the type of `Shapes::idx($s, 'x', shape())` (which is nullable even though we provide a default value because the field is declared as nullable in `Shape`) doesn't satisfy the constraint `T as shape(...)` in the type signature of `Shapes::idx`.

To improve errors like this, I'm proposing to introduce a new reason that will be attached to fake shape types constructed to type check calls to indexing functions.  I also propose to attach the error arising from a mismatch between this fake shape type and the type of the first argument (the indexed shape) to the expression representing the first argument.  At the moment it is attached to the field name, but this doesn't make much sense.  (Note that the above error message points to `'y'`.)

Reviewed By: andrewjkennedy

Differential Revision: D15335006

fbshipit-source-id: 07a43e30f036c98c15c973fbd5e0959ab1dbe35b