Use cached classname in the class type constant to optimize type_structure(cls,cns)['classname']
Summary: In most cases, when people use the function `type_structure()`, they immediately retrieve back `classname` field out. Lets optimize this by caching the name of the class. We convert `type_structure(cls,cns)['classname']` to `type_structure_classname(cls,cns)` to detect this pattern in the JIT. This diff adds IR instructions to check if a classname is cached on a constant and if so, retrieves it. Otherwise, uses a slightly slower path.

Reviewed By: markw65

Differential Revision: D15305287

fbshipit-source-id: a6952da57160b6633d74d700111cf85b88bb2497