Don't produce additional errors when type checking $c::f() where $c has a type error
Summary:
Consider the following example:
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

class C {
  public static function f(): void {}
}

function test(bool $b, mixed $x): void {
  if ($b) {
    /* HH_FIXME[4064] */
    $c = $x->p;
  } else {
    $c = C::class;
  }
  $c::f();
}
```
Right after the if-statement the type of `$c` is inferred to be  `(err | classname<C>)`, and when we type check the call `$c::f()`, we effectively distribute over the union and separately type check that expression with `$c : err` and `$c : classname<C>`.  The latter succeeds while the former produces an error "Was expecting a class but got a type error".  We would like to avoid this kind of cascading errors: we already know there is an error in `$c`, so naturally expressions containing `$c` will also (often) be invalid.

Reviewed By: andrewjkennedy

Differential Revision: D15353901

fbshipit-source-id: db4d3725dce5e889d72c91ccab1d71cffa890efd