Remove hasInstanceDtor() check out of ObjectData::release() fast path
Summary:
Profiling has revealed that the "unlikely" hasInstanceDtor() check in
ObjectData::release() is actually taken a good percentage of the time
(and is responsible for a lot of time in the function). Refactor the
object release logic a bit to avoid this check.

Instead of always calling ObjectData::release() and letting it call
any instance dtor, instead have the Class store a pointer to the
"object release" function. When releasing an object, just do an
indirect call to the release function stored in the class. Normal
classes will just point to ObjectData:release(), but classes with an
instance dtor will point to the dtor function.

As before, when releasing from the TC if we know the exact object type
we can burn in a direct call to the appropriate release function. For
the cases where we don't know the exact type (but we do know its an
object), add a new call type to CallSpec to handle calling the release
function indirectly. The exact case is actually more efficient than
before because we can materialize the Class* as a constant rather than
loading it in ObjectData::release().

Reviewed By: paulbiss

Differential Revision: D15362289

fbshipit-source-id: f702feabac1c0cd7c2f91eb4d1b0afe7b792273a