there should be no need for typing_unify_recursive in new inference
Summary: While working on loop fix points, I found a bug in Typing_unify_recursive which turned out more difficult than expected to fix, so I decided to kill it (at least for new inference), since it's not actually fully required, since we check for bounds containing the type variable before binding. After some adjustments in the previous diffs of this stack, this diff does not add new errors in the code base.

Reviewed By: andrewjkennedy

Differential Revision: D15334892

fbshipit-source-id: 2bc0509c7f7ec9e333ee560d0ba6a0a3719d6cb3