Better expand sets for alias analysis
Summary: We didn't seem to be using the complete versions of these sets anyway, and with this change we can extract more complete information from `expand()` when analyzing kill alias classes. Didn't see a noticeable perf win from this alone but it does seem to improve codegen in some of the examples I looked at.

Reviewed By: mxw

Differential Revision: D15368609

fbshipit-source-id: 53bd4961da4b05e25b1029db8bb77cbdf3deb51f