Add file hash information to sqlite
Summary:
Based on discussion with waleed, add file hash information to sqlite.

The hash is the `SharedMem.get_hash` (an int64) of the relative portion of the filename compared to the root.

With this, we can ensure that if a file is removed off disk, we have a way of avoiding these values in symbol search.

Reviewed By: pittsw

Differential Revision: D15340234

fbshipit-source-id: 58251dbf4c5af53334881f7638e6f6f12d127556