Add tombstones to local search
Summary:
This diff captures the tombstone ID numbers for files that have been removed locally.

When a file is removed locally, the filename is turned into a relative path suffix (relative to the root) and converted into an Int64 hash using `SharedMem.get_hash`.  That tombstone is stored in an `Sset.t` for tombstones.

When a file is updated locally, we remove that hash from the tombstone set.

The intention is for a future diff in this stack to compare values coming out of sqlite and exclude results coming from files in the tombstone set.

I also used this opportunity to remove the other search providers I had experimented with that did not make the cut (grep, ripgrep, etc)

Reviewed By: arxanas

Differential Revision: D15359448

fbshipit-source-id: fb63a805f0434900f50e0cf246539898bdfa7159