Add namespace information to sqlite
Summary:
It's going to become necessary to know what symbols live within which namespaces.

Let's start by keeping a list of known namespaces and add a reference to the namespace name from the symbol.  We'll keep the full namespace in the symbol for now, but we could potentially extract it later and make namespaces fully separate.

Reviewed By: pittsw

Differential Revision: D15339658

fbshipit-source-id: b3f0727d25c4d21481d2b1b52ace6c5a1e9679b1