Load sqlite saved state from watchman/manifold
Summary:
To fetch sqlite saved states, we have to do a few steps:
* Query watchman to find the correct saved state ( thanks to Katie Coons who wrote this feature!)
* Call Manifold to fetch the file
* Call zstd to unzip the file
* Pass the filename to the sqlite search index

This diff implements this behavior.  BUT ... it has one major caveat: When you run in debug mode, the hh server's sha1 signature is not populated, which means this code can never find a legitimately matching saved-state until it is run using an `hh_server` executable that was compiled via the release process.

Bummer!

To make it possible to test the code, I have included a default project metadata key in the code so I can prove that everything else works assuming that it's possible to get a sha1 hash from the release process.  I've also added some log statements to help me determine if all the code works correctly, but still, I guess some tests will have to happen after shipping.

Reviewed By: arxanas

Differential Revision: D15324805

fbshipit-source-id: d0ff7c40a59772659dfc415f325cc395ed0f42d9