Add *_nast functions to Ast provider
Summary: Add functionality to `Ast_provider` to return named ASTs instead of legacy ASTs. This is useful for the unified AST project where we want to remove all usage of the legacy AST. As the `Ast_provider` is used in a few places, this change lets us gradually update individual modules instead of updating everything that depends on the `Ast_provider` at once.

Reviewed By: Wilfred

Differential Revision: D15334183

fbshipit-source-id: 4ac702d2e61c25a47b2622b4560cf422eb30579e