Add an interface for Typing_phase
Summary:
This diff adds a .mli file for `Typing_phase`. This allows us to change the implementation of Typing_phase and identify the parts of the codebase that use it. This is useful for separating external calls to Typing_phase and internal mutually recursive calls. Having this barrier allows you to quickly find all references to a `Typing_phase` function outside Typing_phase itself.

It also means that consumers of localization don't have to worry about the trusted-ness of their types, as Typing_phase can handle it as needed. For example, I found that `localize_hint` (`hint -> locl ty`), while looking similar to other chains of `Decl_hint.hint` (`hint -> decl ty`) and `localize` (`decl ty -> locl ty`), is used for *is* and *as* expressions where the type is required to be trusted, because it is what is passed to the runtime.

Reviewed By: Matt-Schellhas

Differential Revision: D15245956

fbshipit-source-id: 19509eca250f5094db3f02e1fd6ed5b1bef0bec0