Convert exceptions from FFP errors into a Parser error.
Summary:
Today if we throw an exception during `full_fidelity_parser_errors.ml` we will eat the error and report `No Errors` in typechecking and log ```Warning, lowering failed for ... - error: (Failure xxx)``` to the log file no one looks at. This is deeply troubling on many levels.

This diff forces these errors to be reported as `UNEXPECTED_ERROR` with the text of the exception. It won't be very helpful because we don't have a node to report it on, but at least it won't report no errors.

Reviewed By: Matt-Schellhas

Differential Revision: D15397439

fbshipit-source-id: 1fd5366e4a3bfdabe0fa84af7f0bee6035c5acb4