Delete Typing_suggest/Typing_infer_return
Summary: Typing_suggest is an ancient module introduced when Hack first open source. It was meant to infer types in a PHP codebase as a way of boot strapping of the type checker. This isn't needed anymore since Hack no longer is concerned with PHP. I'm also not certain if the feature even works anymore.

Differential Revision: D15404622

fbshipit-source-id: d8f11114d0dcb1c963cfcd632fa18eb73d09a85a