Ban using await on the right of a pipe operator
Summary: Initially I had some complex logic to allow await on both sides of a pipe operator, but due to the way lifting is implemented, this didn't really ever work properly (would give bogus complaint about $$ outside of |> expression). I'm just going to ban it for now and we can re-evaluate allowing it in the future.

Reviewed By: vladima

Differential Revision: D15395949

fbshipit-source-id: d73f6c9eb98ffffd57853e167ec5bce27684d710