Add missing lifting to using statements
Summary: In the process of stripping the gating (D15296659), I realized that `await` lifting inside `using` was missing :O When I added it, it resulting in hack errors due to the analysis that prevents un-`using`-ed disposables. To fix this, I added some logic to prevent lifting the top expression of the using blocks.

Reviewed By: Matt-Schellhas

Differential Revision: D15378745

fbshipit-source-id: 67c7ccdd3462e7ac40295c2ec1848bec06a11ae0