Remove support for PHP4 constructors
Summary:
These are already disallowed by Hack typechecker, although the typechecker does
not produce an error when methods named the same as the class name come via
trait.

Static analysis or PHP7.DeprecationWarnings enabled on an earlier version of
HHVM can be used to discover PHP4 constructors in your codebase.

Reviewed By: oulgen

Differential Revision: D15410243

fbshipit-source-id: 22d122d4c1282016d0f524132e8189da0bc46916