Improve analysis of new static()
Summary:
If `new static()` is called from class B that doesn't have any `__construct()`
methods in its subclass hierarchy, but the `__construct()` method was defined
in its parent class A with overrides outside of B, we failed to resolve the
constructor call to `A::__construct()`.

This is because `cinfo->ctor` pointed to MethTabEntryPair on the class A
implementing the constructor rather than to the flattened entry on B.
Presumably due to complexity of support for PHP4 constructors.

Take advantage of removal of support for PHP4 constructors, simplify the logic
by removing the `info->ctor` field and lookup the constructor's flattened entry
(which contains AttrNoOverride) in `resolve_ctor()`.

Reviewed By: markw65

Differential Revision: D15410533

fbshipit-source-id: d4cb7817ed73c0be6ac9a22603467ca5f2217103