Remove id list from dependent_type
Summary: When I made Typing_defs.dependent_types I kept the list of ids in the path. However we eventually represented these paths as generics instead and not as dependent_types. This means we no longer need these list of ids as part of dependent_types allowing us to simplify this code.

Reviewed By: Wilfred

Differential Revision: D15411299

fbshipit-source-id: e3d2b05529fd4a8404940521689dc09c47f42727