Allow more instruction types in dce
Summary:
We weren't allowing consumesReference instructions, because we didn't
know what to do with their inputs. In the case of PRc instructions, we
can now deal with them correctly, so lets allow some of those.

It turns out that many of them can also raise exceptions, which we
also didn't allow in dce. For things like string concatenation, which
can only raise in oom, or string too long situations, it seems
reasonable to allow such instructions to be eliminated.

Reviewed By: ricklavoie, paulbiss

Differential Revision: D15389204

fbshipit-source-id: e6225015309a501fbaa725272c28f46012fe464f