Differentiate mayMoveReference from consumesReference
Summary:
consumesReference covers two different situations. Either the
instruction decRefs its input, or it transfers ownership of the
input.

We already have movesReference which tells us that it definitely
transfers ownership; but that is conservative, so that new
instructions won't do the wrong thing (refcount opts use this to
deterimine that the input's refcount is unchanged until the next
untracked operation that affects refcounts).

This adds a variant that is conservative in the other direction; if it
returns false, the input is definitely decReffed. This allows us to
place decrefs more optimally when dce kills a consumesRefcount
instruction.

Reviewed By: paulbiss

Differential Revision: D15407544

fbshipit-source-id: 1dc5c9c352e8420c223e5826d55996d86b759527