Do not drop unions as upper bounds on generic type parameters
Summary:
There are two problems with definitions such as
```
function expectLikeNum(~num $x):void { }
function foo<T as ~int>(T $x): void {
  expectLikeNum($x);
}
```
First, the bound is never added to `T` in `tpenv`, because the decomposition code for the subtype constraint `T as ~int` uses the union logic in `simplify_subtype` that currently produces the disjunction `T <: dynamic \/ T <: int`. Fix for this is to introduce special cases for type parameters on left and right when we're using `simplify_subtype` for this purpose.

Second, even when the bound is added to `T`, we solve `T <: ~num` (for example), by doing `T <: dynamic \/ T <: num`, forgetting the possibility `~int <: ~num`. The fix for this is to add an extra case to the disjunction that is constructed when dealing with unions on the right. This also applies to other types with explicit upper bounds (new types, and dependent types).

Reviewed By: manzyuk

Differential Revision: D15431022

fbshipit-source-id: 2c1cd3e960945b233af56e741420bb1c0781ed98